Fixed bug: when flash size was >32kB (ATmega64 and ATmega2560) it didn't work Added support for devices >64kB (ATmega2560) in bluec mode, use set extended address (not tested, yet) Changed flash address and size data types to typedefs Optimized the code size to still fit in ATmega88

+++ bluecontroller/examples/BlueC_ISP/HardwareSerialOpt.cpp
@@ -66,7 +66,7 @@
 
 // avoid 16-bit division for modulo and replace it by AND operation (only possible for power of two buffer sizes)
 #if(RX_BUFFER_SIZE == 0x10 || RX_BUFFER_SIZE == 0x20 || RX_BUFFER_SIZE == 0x40 || RX_BUFFER_SIZE == 0x80 || RX_BUFFER_SIZE == 0x100 || RX_BUFFER_SIZE == 0x200)
-#define MODULO_BUFSIZE(n) (((uint16_t)n) & (RX_BUFFER_SIZE - 1))
+#define MODULO_BUFSIZE(n) (((uint16_t)n) & ((uint16_t)(RX_BUFFER_SIZE - 1)))
 #else
 // TODO: doesn't work, don't know why...
 #error "use a power of two as buffer size..."
